+ [CMN] Common functions definition file. + [CMN] Ring buffer implementation; + [HAL] LED driver; * [HAL] GPIO driver update; * [HAL] GPIO configuration is now a property of platform declaration file; * [HAL] IIR filter for trimmer; * [CMN] Size type is now an unsigned int; * [CMN] Status macro rename; * [PRJ] HAL_Driver updated to version 1.4.1; * [PRJ] USB_Device_Library updated to version 2.4.1;

+++ inc/typedefs.h
@@ -31,7 +31,7 @@
 #else
     typedef ptrdiff_t           Ptr;
 #endif //ptrdiff_t
-    typedef size_t              Size;
+    typedef unsigned int        Size;
 
     typedef float               Float;
     typedef double              Double;
